smart_home_app/mcp/requirements.txt
requests==2.32.3
python-dotenv==1.0.0
mcp==1.8.1
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
aiofiles==24.1.0